+++ web/src/hooks/useMakeAttempt.tsx
@@ -1,8 +1,10 @@
 import { useState } from 'react'
 
 import { useAuth } from '@redwoodjs/auth'
-import { navigate, routes, useParams } from '@redwoodjs/router'
+import { useParams } from '@redwoodjs/router'
 import { useMutation } from '@redwoodjs/web'
+
+import { createAnonAttempt, redirectUser } from 'src/lib/attempt/makeAttempt'
 
 import {
   MakeAttemptMutation,
@@ -115,59 +117,3 @@ const useMakeAttempt = () => {
 }
 
 export default useMakeAttempt
-
-const redirectUser = ({
-  finalStep,
-  isAnon,
-  stepParam,
-  slug,
-}: {
-  finalStep: boolean
-  isAnon: boolean
-  stepParam: string
-  slug: string
-}) => {
-  if (finalStep) {
-    return isAnon
-      ? navigate(routes.anonPuzzleLanding({ slug }))
-      : navigate(routes.puzzleLanding({ slug }))
-  } else {
-    return isAnon
-      ? navigate(
-          routes.anonPuzzleStep({
-            slug,
-            step: parseInt(stepParam, 10) + 1,
-          })
-        )
-      : navigate(
-          routes.puzzleStep({
-            slug,
-            step: parseInt(stepParam, 10) + 1,
-          })
-        )
-  }
-}
-
-const createAnonAttempt = async ({ stepId, stepParam, puzzleId, reqBody }) => {
-  const apiUrl = new URL(
-    `${global.RWJS_API_URL}/anonAttempt`,
-    window.location.origin
-  )
-
-  apiUrl.searchParams.set('stepId', stepId)
-  apiUrl.searchParams.set('stepParam', stepParam)
-  apiUrl.searchParams.set('puzzleId', puzzleId)
-
-  const body = JSON.stringify({ attempt: reqBody })
-
-  try {
-    const response = await fetch(apiUrl, {
-      method: 'POST',
-      body,
-    })
-    const data = await response.json()
-    return data
-  } catch (e) {
-    console.error(e)
-  }
-}
